Tangible



See merge request !1

+++ sloika/networks.py
@@ -1,5 +1,5 @@
 from functools import partial
-from sloika import layers
+from sloika import layers, sloika_dtype
 from numpy.random import normal as rn
 
 
@@ -7,7 +7,7 @@ _NBASE = 4
 _NFEATURE = 4
 
 def _rn(size, sd):
-    return rn(size=size, scale=sd).astype(layers.dtype)
+    return rn(size=size, scale=sd).astype(sloika_dtype)
 
 def _wrap_nanonet(net_layers, kmer, winlen, size, bad_state, sd):
     """ Standard wrappinfg for Nanonet like networks
